DCMAW-12496: Move assertions and test steps
